Update: added jwt authorization validation

+++ src/controllers/Profiles.js
@@ -1,4 +1,5 @@
 import repository from '../repositories/ProfileRepository.js';
+import auth from '../utils/auth.js';
 
 //Get all searchable profiles
 export const getAllProfiles = async (req, res) => {
@@ -22,10 +23,14 @@ export const getProfileById = async (req, res) => {
 
 export const updateProfile = async (req, res) => {
   try {
-    await repository.updateProfile(req.body, req.params.id);
-    res.json({
-      message: 'Perfil atualizado.'
-    });
+    const userId = req.body.userId;
+    if (userId && repository.countProfileByUserId(userId)) {
+      auth.checkToken(userId, req.headers['x-acess-token']);
+      await repository.updateProfile(req.body, req.params.id);
+      res.json({
+        message: 'Perfil atualizado.'
+      });
+    } else throw new Error('Acesso não autorizado.');
   } catch (error) {
     res.json({ message: error.message });
   }
@@ -33,6 +38,7 @@ export const updateProfile = async (req, res) => {
 
 export const createProfile = async (req, res) => {
   try {
+    auth.checkToken(req.body.userId, req.headers['x-acess-token']);
     await repository.createProfile(req.body);
     res.json({
       message: 'Perfil criado.'
@@ -44,10 +50,14 @@ export const createProfile = async (req, res) => {
 
 export const deleteProfile = async (req, res) => {
   try {
-    await repository.deleteProfile(req.params.id);
-    res.json({
-      message: 'Perfil deletado.'
-    });
+    const userId = req.headers['user-id'];
+    if (userId && repository.countProfileByUserId(userId)) {
+      auth.checkToken(userId, req.headers['x-acess-token']);
+      await repository.deleteProfile(req.params.id);
+      res.json({
+        message: 'Perfil deletado.'
+      });
+    } else throw new Error('Acesso não autorizado.');
   } catch (error) {
     res.json({ message: error.message });
   }
